kociemba: Use pruning tables in phase 0
Pruning tables speed up the DFS by pruning, or cutting, the search short
when the number of moves remaining is greater than the amount of moves
needed to solve the current coordinate.

Phase0Tables::prune_depth returns the minimum number of moves needed to
solve the given coordinate. This is the maximum of the number of moves
to solve CO, EO or UD1 separately.